refactor(n6/common): move remote agent types from process/ to common/
RemoteAgentConfig / RemoteAgentInput are consumed by both the renderer
(RemoteAgentManagement.tsx) and the adapter (ipcBridge.ts type imports),
so the canonical home is common/types/ rather than process/. With the
move, process/agent/ becomes empty and is removed.

- Add: common/types/remoteAgentTypes.ts (51 lines, identical content)
- Delete: process/agent/remote/types.ts
- Delete: empty directories process/agent/remote/, process/agent/
- Update 6 import sites (1 tsx + 5 inline type imports in ipcBridge.ts)

Audit sections 4.7.1 / 7.3 / 6.2 C9.
